# Adding sparse histogram support