--- a/engine/engine.go
+++ b/engine/engine.go
@@ -370,7 +370,6 @@ loop:
 				for i := range r {
 					numSeries += len(r[i].Samples)
 				}
-
 				series = make([]promql.Series, numSeries)
 
 				for _, vector := range r {
@@ -395,18 +394,16 @@ loop:
 						T: vector.T,
 						V: vector.Samples[i],
 					})
-
 				}
-				for i, s := range vector.HistogramSampleIDs {
+				for i, s := range vector.HistogramIDs {
 					if len(series[s].Points) == 0 {
 						series[s].Points = make([]promql.Point, 0, 121) // Typically 1h of data.
 					}
 					series[s].Points = append(series[s].Points, promql.Point{
 						T: vector.T,
-						H: vector.HistogramSamples[i],
+						H: vector.Histograms[i],
 					})
 				}
-
 				q.Query.exec.GetPool().PutStepVector(vector)
 			}
 			q.Query.exec.GetPool().PutVectors(r)