--- a/execution/function/functions.go
+++ b/execution/function/functions.go
@@ -8,6 +8,7 @@ import (
 	"math"
 
 	"github.com/efficientgo/core/errors"
+	"github.com/prometheus/prometheus/model/histogram"
 	"github.com/prometheus/prometheus/model/labels"
 	"github.com/prometheus/prometheus/promql"
 	"github.com/prometheus/prometheus/promql/parser"
@@ -301,11 +302,13 @@ var Funcs = map[string]FunctionCall{
 		if len(f.Points) < 2 {
 			return InvalidSample
 		}
+		v, h := extrapolatedRate(f.Points, true, true, f.StepTime, f.SelectRange, f.Offset)
 		return promql.Sample{
 			Metric: f.Labels,
 			Point: promql.Point{
 				T: f.StepTime,
-				V: extrapolatedRate(f.Points, true, true, f.StepTime, f.SelectRange, f.Offset),
+				V: v,
+				H: h,
 			},
 		}
 	},
@@ -313,11 +316,13 @@ var Funcs = map[string]FunctionCall{
 		if len(f.Points) < 2 {
 			return InvalidSample
 		}
+		v, h := extrapolatedRate(f.Points, true, true, f.StepTime, f.SelectRange, f.Offset)
 		return promql.Sample{
 			Metric: f.Labels,
 			Point: promql.Point{
 				T: f.StepTime,
-				V: extrapolatedRate(f.Points, false, false, f.StepTime, f.SelectRange, f.Offset),
+				V: v,
+				H: h,
 			},
 		}
 	},
@@ -325,11 +330,13 @@ var Funcs = map[string]FunctionCall{
 		if len(f.Points) < 2 {
 			return InvalidSample
 		}
+		v, h := extrapolatedRate(f.Points, true, true, f.StepTime, f.SelectRange, f.Offset)
 		return promql.Sample{
 			Metric: f.Labels,
 			Point: promql.Point{
 				T: f.StepTime,
-				V: extrapolatedRate(f.Points, true, false, f.StepTime, f.SelectRange, f.Offset),
+				V: v,
+				H: h,
 			},
 		}
 	},
@@ -405,20 +412,26 @@ func NewFunctionCall(f *parser.Function) (FunctionCall, error) {
 // It calculates the rate (allowing for counter resets if isCounter is true),
 // extrapolates if the first/last sample is close to the boundary, and returns
 // the result as either per-second (if isRate is true) or overall.
-func extrapolatedRate(samples []promql.Point, isCounter, isRate bool, stepTime int64, selectRange int64, offset int64) float64 {
+func extrapolatedRate(samples []promql.Point, isCounter, isRate bool, stepTime int64, selectRange int64, offset int64) (float64, *histogram.FloatHistogram) {
 	var (
-		rangeStart = stepTime - (selectRange + offset)
-		rangeEnd   = stepTime - offset
+		rangeStart      = stepTime - (selectRange + offset)
+		rangeEnd        = stepTime - offset
+		resultValue     float64
+		resultHistogram *histogram.FloatHistogram
 	)
 
-	resultValue := samples[len(samples)-1].V - samples[0].V
-	if isCounter {
-		var lastValue float64
-		for _, sample := range samples {
-			if sample.V < lastValue {
-				resultValue += lastValue
+	if samples[0].H != nil {
+		resultHistogram = histogramRate(samples, isCounter)
+	} else {
+		resultValue := samples[len(samples)-1].V - samples[0].V
+		if isCounter {
+			var lastValue float64
+			for _, sample := range samples {
+				if sample.V < lastValue {
+					resultValue += lastValue
+				}
+				lastValue = sample.V
 			}
-			lastValue = sample.V
 		}
 	}
 
@@ -464,9 +477,62 @@ func extrapolatedRate(samples []promql.Point, isCounter, isRate bool, stepTime i
 	if isRate {
 		factor /= float64(selectRange / 1000)
 	}
-	resultValue *= factor
+	if resultHistogram == nil {
+		resultValue *= factor
+	} else {
+		resultHistogram.Scale(factor)
 
-	return resultValue
+	}
+
+	return resultValue, resultHistogram
+}
+
+// histogramRate is a helper function for extrapolatedRate. It requires
+// points[0] to be a histogram. It returns nil if any other Point in points is
+// not a histogram.
+func histogramRate(points []promql.Point, isCounter bool) *histogram.FloatHistogram {
+	prev := points[0].H // We already know that this is a histogram.
+	last := points[len(points)-1].H
+	if last == nil {
+		return nil // Range contains a mix of histograms and floats.
+	}
+	minSchema := prev.Schema
+	if last.Schema < minSchema {
+		minSchema = last.Schema
+	}
+
+	// First iteration to find out two things:
+	// - What's the smallest relevant schema?
+	// - Are all data points histograms?
+	//   TODO(beorn7): Find a way to check that earlier, e.g. by handing in a
+	//   []FloatPoint and a []HistogramPoint separately.
+	for _, currPoint := range points[1 : len(points)-1] {
+		curr := currPoint.H
+		if curr == nil {
+			return nil // Range contains a mix of histograms and floats.
+		}
+		if !isCounter {
+			continue
+		}
+		if curr.Schema < minSchema {
+			minSchema = curr.Schema
+		}
+	}
+
+	h := last.CopyToSchema(minSchema)
+	h.Sub(prev)
+
+	if isCounter {
+		// Second iteration to deal with counter resets.
+		for _, currPoint := range points[1:] {
+			curr := currPoint.H
+			if curr.DetectReset(prev) {
+				h.Add(prev)
+			}
+			prev = curr
+		}
+	}
+	return h.Compact(0)
 }
 
 func instantValue(samples []promql.Point, isRate bool) (float64, bool) {