This PR will add following function for sparse histograms

- [x] rate()
- [x] increase()
- [x] delta()



Issues: https://github.com/thanos-community/promql-engine/issues/149