@fpetkovski 2023-01-27
Looks like we're good for now, we can add support for `sum` and `histogram_x` functions in followup PRs.

Benchmark against main branch:

```
name                                                  old time/op    new time/op    delta
RangeQuery/vector_selector-8                            17.0ms ± 5%    15.5ms ±13%     ~     (p=0.095 n=5+5)
RangeQuery/sum-8                                        11.7ms ± 1%    11.0ms ± 1%   -5.91%  (p=0.016 n=4+5)
RangeQuery/sum_by_pod-8                                 21.5ms ±28%    15.9ms ±14%  -25.95%  (p=0.016 n=5+5)
RangeQuery/topk-8                                       14.0ms ±22%    11.9ms ± 8%     ~     (p=0.222 n=5+5)
RangeQuery/bottomk-8                                    13.9ms ±24%    11.1ms ± 2%  -20.43%  (p=0.008 n=5+5)
RangeQuery/rate-8                                       23.3ms ±18%    22.4ms ±14%     ~     (p=0.310 n=5+5)
RangeQuery/sum_rate-8                                   21.6ms ±15%    18.9ms ± 2%  -12.33%  (p=0.008 n=5+5)
RangeQuery/sum_by_rate-8                                24.7ms ±10%    23.5ms ± 8%     ~     (p=0.222 n=5+5)
RangeQuery/quantile_with_variable_parameter-8           37.2ms ±27%    27.3ms ± 6%  -26.69%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_one_to_one-8           14.0ms ± 5%    13.9ms ± 2%     ~     (p=0.548 n=5+5)
RangeQuery/binary_operation_with_many_to_one-8          29.9ms ± 2%    31.2ms ± 5%     ~     (p=0.222 n=5+5)
RangeQuery/binary_operation_with_vector_and_scalar-8    19.0ms ± 2%    21.1ms ±12%  +10.68%  (p=0.008 n=5+5)
RangeQuery/unary_negation-8                             15.9ms ± 4%    17.3ms ± 9%   +8.96%  (p=0.032 n=5+5)
RangeQuery/vector_and_scalar_comparison-8               19.7ms ± 1%    20.8ms ± 1%   +5.29%  (p=0.008 n=5+5)
RangeQuery/positive_offset_vector-8                     13.2ms ± 0%    14.4ms ±21%   +9.24%  (p=0.008 n=5+5)
RangeQuery/at_modifier_-8                               10.2ms ± 4%    10.3ms ± 6%     ~     (p=0.841 n=5+5)
RangeQuery/at_modifier_with_positive_offset_vector-8    10.2ms ± 7%     9.8ms ± 2%     ~     (p=0.095 n=5+5)
RangeQuery/clamp-8                                      30.2ms ± 9%    27.6ms ± 0%     ~     (p=0.095 n=5+5)
RangeQuery/clamp_min-8                                  25.3ms ± 2%    24.3ms ± 3%     ~     (p=0.056 n=5+5)
RangeQuery/complex_func_query-8                         35.6ms ± 5%    34.7ms ± 3%     ~     (p=0.548 n=5+5)
RangeQuery/func_within_func_query-8                     29.1ms ± 6%    28.8ms ± 4%     ~     (p=0.421 n=5+5)
RangeQuery/aggr_within_func_query-8                     37.6ms ±23%    28.2ms ± 2%  -25.01%  (p=0.008 n=5+5)
RangeQuery/histogram_quantile-8                         97.0ms ± 9%    96.0ms ± 1%     ~     (p=0.548 n=5+5)

name                                                  old alloc/op   new alloc/op   delta
RangeQuery/vector_selector-8                            39.1MB ± 0%    39.1MB ± 0%     ~     (p=0.056 n=5+5)
RangeQuery/sum-8                                        10.7MB ± 2%    10.3MB ± 0%   -3.77%  (p=0.008 n=5+5)
RangeQuery/sum_by_pod-8                                 23.6MB ± 0%    23.7MB ± 0%   +0.47%  (p=0.008 n=5+5)
RangeQuery/topk-8                                       12.3MB ± 3%    12.0MB ± 2%     ~     (p=0.310 n=5+5)
RangeQuery/bottomk-8                                    12.2MB ± 1%    11.5MB ± 2%   -5.93%  (p=0.008 n=5+5)
RangeQuery/rate-8                                       41.8MB ± 0%    41.8MB ± 0%     ~     (p=0.548 n=5+5)
RangeQuery/sum_rate-8                                   13.7MB ± 2%    13.1MB ± 1%   -4.72%  (p=0.008 n=5+5)
RangeQuery/sum_by_rate-8                                26.1MB ± 1%    25.8MB ± 1%     ~     (p=0.056 n=5+5)
RangeQuery/quantile_with_variable_parameter-8           41.2MB ± 0%    41.5MB ± 0%   +0.77%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_one_to_one-8           22.5MB ± 0%    22.7MB ± 0%   +0.81%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_many_to_one-8          57.2MB ± 0%    57.4MB ± 0%   +0.32%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_vector_and_scalar-8    42.0MB ± 0%    42.1MB ± 0%     ~     (p=0.151 n=5+5)
RangeQuery/unary_negation-8                             40.4MB ± 0%    40.5MB ± 0%   +0.31%  (p=0.008 n=5+5)
RangeQuery/vector_and_scalar_comparison-8               41.7MB ± 0%    41.8MB ± 0%     ~     (p=0.095 n=5+5)
RangeQuery/positive_offset_vector-8                     37.2MB ± 0%    37.3MB ± 0%   +0.23%  (p=0.008 n=5+5)
RangeQuery/at_modifier_-8                               33.2MB ± 0%    33.2MB ± 0%   +0.18%  (p=0.008 n=5+5)
RangeQuery/at_modifier_with_positive_offset_vector-8    33.0MB ± 0%    33.0MB ± 0%   +0.18%  (p=0.008 n=5+5)
RangeQuery/clamp-8                                      39.4MB ± 0%    39.5MB ± 0%   +0.16%  (p=0.008 n=5+5)
RangeQuery/clamp_min-8                                  39.4MB ± 0%    39.5MB ± 0%   +0.12%  (p=0.032 n=5+5)
RangeQuery/complex_func_query-8                         42.4MB ± 0%    42.3MB ± 0%     ~     (p=0.222 n=5+5)
RangeQuery/func_within_func_query-8                     42.0MB ± 0%    42.1MB ± 0%   +0.22%  (p=0.008 n=5+5)
RangeQuery/aggr_within_func_query-8                     42.0MB ± 0%    42.0MB ± 0%   +0.15%  (p=0.008 n=5+5)
RangeQuery/histogram_quantile-8                         61.8MB ± 0%    61.9MB ± 0%     ~     (p=0.421 n=5+5)

name                                                  old allocs/op  new allocs/op  delta
RangeQuery/vector_selector-8                              129k ± 0%      129k ± 0%   -0.34%  (p=0.008 n=5+5)
RangeQuery/sum-8                                          124k ± 0%      123k ± 0%   -0.47%  (p=0.008 n=5+5)
RangeQuery/sum_by_pod-8                                   209k ± 0%      209k ± 0%   -0.17%  (p=0.008 n=5+5)
RangeQuery/topk-8                                         129k ± 0%      132k ± 0%   +2.24%  (p=0.008 n=5+5)
RangeQuery/bottomk-8                                      132k ± 0%      131k ± 0%   -1.03%  (p=0.008 n=5+5)
RangeQuery/rate-8                                         153k ± 0%      153k ± 0%   -0.28%  (p=0.008 n=5+5)
RangeQuery/sum_rate-8                                     148k ± 0%      147k ± 0%   -0.47%  (p=0.008 n=5+5)
RangeQuery/sum_by_rate-8                                  233k ± 0%      232k ± 0%   -0.24%  (p=0.008 n=5+5)
RangeQuery/quantile_with_variable_parameter-8             614k ± 0%      613k ± 0%   -0.11%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_one_to_one-8             103k ± 0%      103k ± 0%   -0.22%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_many_to_one-8            197k ± 0%      197k ± 0%   -0.17%  (p=0.008 n=5+5)
RangeQuery/binary_operation_with_vector_and_scalar-8      133k ± 0%      133k ± 0%   -0.25%  (p=0.008 n=5+5)
RangeQuery/unary_negation-8                               142k ± 0%      141k ± 0%   -0.24%  (p=0.008 n=5+5)
RangeQuery/vector_and_scalar_comparison-8                 130k ± 0%      130k ± 0%   -0.28%  (p=0.008 n=5+5)
RangeQuery/positive_offset_vector-8                       100k ± 0%      100k ± 0%   -0.25%  (p=0.008 n=5+5)
RangeQuery/at_modifier_-8                                74.8k ± 0%     74.7k ± 0%   -0.15%  (p=0.008 n=5+5)
RangeQuery/at_modifier_with_positive_offset_vector-8     68.8k ± 0%     68.7k ± 0%   -0.17%  (p=0.008 n=5+5)
RangeQuery/clamp-8                                        133k ± 0%      133k ± 0%   -0.32%  (p=0.008 n=5+5)
RangeQuery/clamp_min-8                                    133k ± 0%      132k ± 0%   -0.30%  (p=0.008 n=5+5)
RangeQuery/complex_func_query-8                           137k ± 0%      137k ± 0%   -0.34%  (p=0.008 n=5+5)
RangeQuery/func_within_func_query-8                       154k ± 0%      154k ± 0%   -0.25%  (p=0.008 n=5+5)
RangeQuery/aggr_within_func_query-8                       154k ± 0%      154k ± 0%   -0.27%  (p=0.008 n=5+5)
RangeQuery/histogram_quantile-8                           713k ± 0%      713k ± 0%   +0.01%  (p=0.016 n=5+5)
```